ChemistryCodeToJSON now accepts arguments: ChemistryToJSON [type!(atoms/molecules)] [title?] [version?] [creator?]
